% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.R
\name{df_to_gneiss}
\alias{df_to_gneiss}
\title{Convert data.frame to gneiss Format}
\usage{
df_to_gneiss(data, label, type, data_source)
}
\arguments{
\item{data}{data.frame}

\item{label}{column name of labels}

\item{type}{chart types}

\item{data_source}{source of the data}
}
\description{
Convert data.frame to gneiss Format
}
\seealso{
\link{gneiss}
}

