% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gneiss.R
\name{gneiss}
\alias{gneiss}
\title{Draw A Chart As Quartz Does}
\usage{
gneiss(x, label, type = "line", credit = "Gneisschart",
  data_source = NULL, width = NULL, height = NULL)
}
\description{
An R interface to gneisschart
}

