% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gneisschart.R
\docType{package}
\name{gneisschart}
\alias{gneisschart}
\alias{gneisschart-package}
\title{gneisschart}
\description{
An R package for drawing charts by Quartz's Gneisschart
}
\seealso{
\url{https://github.com/Quartz/Chartbuilder}
}

